Move remaining missed files per Andrew
