+++ src/applications/financial-status-report/components/employment/EmploymentHistoryWidget.jsx
@@ -1,13 +1,13 @@
 import React, { useEffect } from 'react';
-import { useSelector, connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router';
-import EmploymentHistorySummaryCard from '../../../components/employment/EmploymentHistorySummaryCard';
-import { EmptyMiniSummaryCard } from '../../../components/shared/MiniSummaryCard';
+import { useSelector, connect } from 'react-redux';
+import EmploymentHistorySummaryCard from './EmploymentHistorySummaryCard';
+import { EmptyMiniSummaryCard } from '../shared/MiniSummaryCard';
 import FormNavButtons from '~/platform/forms-system/src/js/components/FormNavButtons';
-import { clearJobIndex } from '../../../utils/session';
+import { clearJobIndex } from '../../utils/session';
 
-const SpouseEmploymentHistoryWidget = props => {
+const EmploymentHistoryWidget = props => {
   const {
     goToPath,
     goForward,
@@ -18,8 +18,8 @@ const SpouseEmploymentHistoryWidget = props => {
 
   const formData = useSelector(state => state.form.data);
   const employmentHistory =
-    formData.personalData.employmentHistory.spouse.spEmploymentRecords || [];
-  const efsrFeatureFlag = formData['view:enhancedFinancialStatusReport'];
+    formData.personalData.employmentHistory.veteran.employmentRecords || [];
+
   useEffect(() => {
     clearJobIndex();
   }, []);
@@ -27,7 +27,7 @@ const SpouseEmploymentHistoryWidget = props => {
   const handlers = {
     onBackClick: event => {
       event.preventDefault();
-      goToPath('/enhanced-spouse-employment-question');
+      goToPath('/employment-question');
     },
   };
 
@@ -39,53 +39,45 @@ const SpouseEmploymentHistoryWidget = props => {
     />
   );
   const updateButton = <button type="submit">Review update button</button>;
-  const emptyPrompt = `Select the ‘add additional job link to add another job. Select the continue button to move on to the next question.`;
 
   return (
-    <form>
+    <form onSubmit={handlers.onSubmit}>
       <fieldset className="vads-u-margin-y--2">
         <legend className="schemaform-block-title">
-          <h3 className="vads-u-margin--0">Your spouse’s work history</h3>
+          <h3 className="vads-u-margin--0">Your work history</h3>
         </legend>
         <div className="vads-u-margin-y--3" data-testid="debt-list">
           {employmentHistory.length === 0 ? (
-            <EmptyMiniSummaryCard content={emptyPrompt} />
+            <EmptyMiniSummaryCard content="No employment history provided" />
           ) : (
             employmentHistory.map((job, index) => (
               <EmploymentHistorySummaryCard
                 key={`${index}-${job.employername}`}
                 job={job}
                 index={index}
-                isSpouse
+                isSpouse={false}
               />
             ))
           )}
         </div>
         <Link
           className="vads-c-action-link--green"
-          to={
-            efsrFeatureFlag
-              ? '/enhanced-spouse-employment-records'
-              : '/spouse-employment-records'
-          }
+          to="/enhanced-employment-records"
         >
           Add another job from the last 2 years
         </Link>
       </fieldset>
       {contentBeforeButtons}
       {onReviewPage ? updateButton : navButtons}
-      {contentAfterButtons}
+      {contentAfterButtons}{' '}
     </form>
   );
 };
-
-SpouseEmploymentHistoryWidget.propTypes = {
-  goBack: PropTypes.func.isRequired,
+EmploymentHistoryWidget.propTypes = {
   goForward: PropTypes.func.isRequired,
   goToPath: PropTypes.func.isRequired,
   onReviewPage: PropTypes.bool,
 };
-
 const mapStateToProps = ({ form }) => {
   return {
     formData: form.data,
@@ -93,4 +85,4 @@ const mapStateToProps = ({ form }) => {
   };
 };
 
-export default connect(mapStateToProps)(SpouseEmploymentHistoryWidget);
+export default connect(mapStateToProps)(EmploymentHistoryWidget);
