+++ src/applications/financial-status-report/components/employment/SpouseEmploymentHistoryWidget.jsx
@@ -1,13 +1,13 @@
 import React, { useEffect } from 'react';
+import { useSelector, connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router';
-import { useSelector, connect } from 'react-redux';
-import EmploymentHistorySummaryCard from '../../../components/employment/EmploymentHistorySummaryCard';
-import { EmptyMiniSummaryCard } from '../../../components/shared/MiniSummaryCard';
+import EmploymentHistorySummaryCard from './EmploymentHistorySummaryCard';
+import { EmptyMiniSummaryCard } from '../shared/MiniSummaryCard';
 import FormNavButtons from '~/platform/forms-system/src/js/components/FormNavButtons';
-import { clearJobIndex } from '../../../utils/session';
+import { clearJobIndex } from '../../utils/session';
 
-const EmploymentHistoryWidget = props => {
+const SpouseEmploymentHistoryWidget = props => {
   const {
     goToPath,
     goForward,
@@ -18,8 +18,8 @@ const EmploymentHistoryWidget = props => {
 
   const formData = useSelector(state => state.form.data);
   const employmentHistory =
-    formData.personalData.employmentHistory.veteran.employmentRecords || [];
-
+    formData.personalData.employmentHistory.spouse.spEmploymentRecords || [];
+  const efsrFeatureFlag = formData['view:enhancedFinancialStatusReport'];
   useEffect(() => {
     clearJobIndex();
   }, []);
@@ -27,7 +27,7 @@ const EmploymentHistoryWidget = props => {
   const handlers = {
     onBackClick: event => {
       event.preventDefault();
-      goToPath('/employment-question');
+      goToPath('/enhanced-spouse-employment-question');
     },
   };
 
@@ -39,45 +39,53 @@ const EmploymentHistoryWidget = props => {
     />
   );
   const updateButton = <button type="submit">Review update button</button>;
+  const emptyPrompt = `Select the ‘add additional job link to add another job. Select the continue button to move on to the next question.`;
 
   return (
-    <form onSubmit={handlers.onSubmit}>
+    <form>
       <fieldset className="vads-u-margin-y--2">
         <legend className="schemaform-block-title">
-          <h3 className="vads-u-margin--0">Your work history</h3>
+          <h3 className="vads-u-margin--0">Your spouse’s work history</h3>
         </legend>
         <div className="vads-u-margin-y--3" data-testid="debt-list">
           {employmentHistory.length === 0 ? (
-            <EmptyMiniSummaryCard content="No employment history provided" />
+            <EmptyMiniSummaryCard content={emptyPrompt} />
           ) : (
             employmentHistory.map((job, index) => (
               <EmploymentHistorySummaryCard
                 key={`${index}-${job.employername}`}
                 job={job}
                 index={index}
-                isSpouse={false}
+                isSpouse
               />
             ))
           )}
         </div>
         <Link
           className="vads-c-action-link--green"
-          to="/enhanced-employment-records"
+          to={
+            efsrFeatureFlag
+              ? '/enhanced-spouse-employment-records'
+              : '/spouse-employment-records'
+          }
         >
           Add another job from the last 2 years
         </Link>
       </fieldset>
       {contentBeforeButtons}
       {onReviewPage ? updateButton : navButtons}
-      {contentAfterButtons}{' '}
+      {contentAfterButtons}
     </form>
   );
 };
-EmploymentHistoryWidget.propTypes = {
+
+SpouseEmploymentHistoryWidget.propTypes = {
+  goBack: PropTypes.func.isRequired,
   goForward: PropTypes.func.isRequired,
   goToPath: PropTypes.func.isRequired,
   onReviewPage: PropTypes.bool,
 };
+
 const mapStateToProps = ({ form }) => {
   return {
     formData: form.data,
@@ -85,4 +93,4 @@ const mapStateToProps = ({ form }) => {
   };
 };
 
-export default connect(mapStateToProps)(EmploymentHistoryWidget);
+export default connect(mapStateToProps)(SpouseEmploymentHistoryWidget);
